added apply operator functions
