+++ pkg/tfutil/func.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"math/cmplx"
 
-	tfutilop "github.com/kubetrail/tfutil/pkg/op"
 	tf "github.com/tensorflow/tensorflow/tensorflow/go"
 	"github.com/tensorflow/tensorflow/tensorflow/go/op"
 )
@@ -12,208 +11,22 @@ import (
 // MatrixInverse inverts the tensor assuming it is a square matrix,
 // otherwise it will throw error
 func MatrixInverse[T PrimitiveTypes](input *Tensor[T]) (*Tensor[T], error) {
-	x, err := input.Marshal()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get tf tensor: %w", err)
-	}
-
-	root := op.NewScope()
-	X := op.Placeholder(
-		root.SubScope("X"),
-		x.DataType(),
-		op.PlaceholderShape(
-			tf.MakeShape(castToInt64(input.shape)...),
-		),
+	return ApplyOperators(
+		input,
+		func(scope *op.Scope, x tf.Output) tf.Output {
+			return op.MatrixInverse(scope, x)
+		},
 	)
-
-	// operation to invert matrix.
-	// needs a square matrix
-	Output := tfutilop.MatrixInverse(root, X)
-
-	graph, err := root.Finalize()
-	if err != nil {
-		return nil, fmt.Errorf("failed to import graph: %w", err)
-	}
-
-	feeds := map[tf.Output]*tf.Tensor{
-		X: x,
-	}
-
-	fetches := []tf.Output{
-		Output,
-	}
-
-	sess, err := tf.NewSession(
-		graph,
-		&tf.SessionOptions{},
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create new tf session: %w", err)
-	}
-	defer func(sess *tf.Session) {
-		err := sess.Close()
-		if err != nil {
-			panic(err)
-		}
-	}(sess)
-
-	out, err := sess.Run(feeds, fetches, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to run tf session: %w", err)
-	}
-
-	if len(out) != 1 {
-		return nil, fmt.Errorf("expected session run output to have length 1, got %d", len(out))
-	}
-
-	output := &Tensor[T]{}
-	if err := output.Unmarshal(out[0]); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal output: %w", err)
-	}
-
-	return output, nil
 }
 
 // MatrixMultiply performs matrix multiplication
 func MatrixMultiply[T PrimitiveTypes](x, y *Tensor[T]) (*Tensor[T], error) {
-	xTfTensor, err := x.Marshal()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get tf tensor: %w", err)
-	}
-
-	yTfTensor, err := y.Marshal()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get tf tensor: %w", err)
-	}
-
-	root := op.NewScope()
-	X := op.Placeholder(
-		root.SubScope("X"),
-		xTfTensor.DataType(),
-		op.PlaceholderShape(
-			tf.MakeShape(castToInt64(x.shape)...),
-		),
+	return ApplyOperatorXY(
+		x, y,
+		func(scope *op.Scope, x, y tf.Output) tf.Output {
+			return op.MatMul(scope, x, y)
+		},
 	)
-
-	Y := op.Placeholder(
-		root.SubScope("Y"),
-		yTfTensor.DataType(),
-		op.PlaceholderShape(
-			tf.MakeShape(castToInt64(y.shape)...),
-		),
-	)
-
-	// operation to invert matrix.
-	// needs a square matrix
-	Output := op.MatMul(root, X, Y)
-
-	graph, err := root.Finalize()
-	if err != nil {
-		return nil, fmt.Errorf("failed to import graph: %w", err)
-	}
-
-	feeds := map[tf.Output]*tf.Tensor{
-		X: xTfTensor,
-		Y: yTfTensor,
-	}
-
-	fetches := []tf.Output{
-		Output,
-	}
-
-	sess, err := tf.NewSession(
-		graph,
-		&tf.SessionOptions{},
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create new tf session: %w", err)
-	}
-	defer func(sess *tf.Session) {
-		err := sess.Close()
-		if err != nil {
-			panic(err)
-		}
-	}(sess)
-
-	out, err := sess.Run(feeds, fetches, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to run tf session: %w", err)
-	}
-
-	if len(out) != 1 {
-		return nil, fmt.Errorf("expected session run output to have length 1, got %d", len(out))
-	}
-
-	output := &Tensor[T]{}
-	if err := output.Unmarshal(out[0]); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal output: %w", err)
-	}
-
-	return output, nil
-}
-
-// Round rounds the elements of input matrix
-func Round[T PrimitiveTypes](input *Tensor[T]) (*Tensor[T], error) {
-	x, err := input.Marshal()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get tf tensor: %w", err)
-	}
-
-	root := op.NewScope()
-	X := op.Placeholder(
-		root.SubScope("X"),
-		x.DataType(),
-		op.PlaceholderShape(
-			tf.MakeShape(castToInt64(input.shape)...),
-		),
-	)
-
-	// operation to invert matrix.
-	// needs a square matrix
-	Output := op.Round(root, X)
-
-	graph, err := root.Finalize()
-	if err != nil {
-		return nil, fmt.Errorf("failed to import graph: %w", err)
-	}
-
-	feeds := map[tf.Output]*tf.Tensor{
-		X: x,
-	}
-
-	fetches := []tf.Output{
-		Output,
-	}
-
-	sess, err := tf.NewSession(
-		graph,
-		&tf.SessionOptions{},
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create new tf session: %w", err)
-	}
-	defer func(sess *tf.Session) {
-		err := sess.Close()
-		if err != nil {
-			panic(err)
-		}
-	}(sess)
-
-	out, err := sess.Run(feeds, fetches, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to run tf session: %w", err)
-	}
-
-	if len(out) != 1 {
-		return nil, fmt.Errorf("expected session run output to have length 1, got %d", len(out))
-	}
-
-	output := &Tensor[T]{}
-	if err := output.Unmarshal(out[0]); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal output: %w", err)
-	}
-
-	return output, nil
 }
 
 // Transpose transposes a tensor. perm refers to the new order
